allUniPlots {iNZightModules}	R Documentation

Explore All Univariate Plots

Description

Explore all univariate plots

Usage

allUniPlots(e)

Arguments

e
(depreciated ... will be updated in future)

Details

Opens a pop-up window that allows users to click through all univariate plots, which is useful for quickly exploring large data sets.

Author(s)

Tom Elliott

[Package iNZightModules version 2.3-1 Index]